requirements.txt
accelerate==0.27.2
aiodns==3.4.0
aiohappyeyeballs==2.6.1
aiohttp==3.10.11
aiosignal==1.3.2
alpaca-py==0.40.1
alpaca-trade-api==0.26
annotated-types==0.7.0
asyncio-nats-client==0.11.5
attrs==25.3.0
beautifulsoup4==4.13.4
blinker==1.9.0
ccxt==4.4.82
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.0
cryptography==45.0.2
curl_cffi==0.11.1
data_fetcher==0.1.9
deprecation==2.1.0
filelock==3.18.0
Flask==3.1.1
frozendict==2.4.6
frozenlist==1.6.0
fsspec==2025.5.1
greenlet==3.2.2
hf-xet==1.1.3
huggingface-hub==0.33.0
idna==3.10
iniconfig==2.1.0
itsdangerous==2.2.0
Jinja2==3.1.6
lxml==5.4.0
MarkupSafe==3.0.2
mpmath==1.3.0
msgpack==1.0.3
multidict==6.4.3
multitasking==0.0.11
networkx==3.5
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.19.3
nvidia-nvjitlink-cu12==12.9.86
nvidia-nvtx-cu12==12.1.105
packaging==25.0
pandas==2.2.3
pandas_ta==0.3.14b0
peewee==3.18.1
platformdirs==4.3.8
pluggy==1.6.0
propcache==0.3.1
protobuf==6.31.0
psutil==7.0.0
psycopg2-binary==2.9.10
pycares==4.8.0
pycparser==2.22
pydantic==2.11.4
pydantic_core==2.33.2
Pygments==2.19.1
pytest==8.4.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
PyYAML==6.0.1
regex==2024.11.6
requests==2.32.3
safetensors==0.5.3
sentencepiece==0.2.0
setuptools==80.7.1
six==1.17.0
soupsieve==2.7
SQLAlchemy==2.0.41
sseclient-py==1.8.0
sympy==1.14.0
tenacity==9.1.2
tokenizers==0.15.2
torch==2.2.1
tqdm==4.67.1
transformers==4.38.2
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
urllib3==1.26.20
websocket-client==1.8.0
websockets==15.0.1
Werkzeug==3.1.3
yarl==1.20.0
yfinance==0.2.61
